Move stateful Components to containers
