+++ src/__test__/components/TodayMeals.test.js
@@ -1,6 +1,6 @@
 import { Provider } from 'react-redux';
 import { create } from 'react-test-renderer';
-import TodayMeal from '../../components/todayMeal';
+import TodayMeal from '../../containers/todayMeal';
 import store from '../../redux';
 
 test('renders Correctly', () => {
