+++ src/containers/recentlyViewed/index.js
