+++ src/pages/homePage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Container } from 'react-bootstrap';
-import TodayMeal from '../components/todayMeal';
+import TodayMeal from '../containers/todayMeal';
 import Carousel from '../components/carousel';
 
 export default function HomePage() {
